Make GenServer safe on application start
On start_link, the cache function attempts to use avlizer_confluent().make_encoder().
The issue is this erlang process can crash if it connect connect or
cannot find the schema.
This causes this GenServer to die, which then kills the main application
Here instead we just don't attempt to make requests unless we have a known id.